@@ -1,54 +1,73 @@
-var React = require("react/addons");
+var React = require("react");
+var ReactDOM = require('react-dom');
 var Reflux = require("reflux");
 var Chatty = require("./view/chatty.js");
 var StatusBar = require("./view/user/statusbar.js");
 var StatusMenu = require("./view/user/statusmenu.js");
 var ChattyStore = require("./store/chattystore.js");
 var UserStore = require("./store/userstore.js");
-var Router = require('react-router');
-var { Route, DefaultRoute, RouteHandler, Link } = Router;
+import { Router, Route, IndexRoute } from 'react-router';
 var ChattyHome = require("./routes/chattyhome.js");
 var SinglePost = require("./routes/singlepost.js");
+var ChattyActions = require("./store/chattyactions.js");
+var UserActions = require("./store/useractions.js");
+var keymaster = require("keymaster");
 
 var App = React.createClass({
   mixins: [
     Reflux.connect(UserStore,"UserStore"),
     Reflux.connect(ChattyStore,"ChattyStore"),
   ],
+  componentDidMount: function() {
+    keymaster('a', function() {
+       ChattyActions.selectPrevComment();
+    });
+    keymaster('z', function() {
+        ChattyActions.selectNextComment();
+    });
+  },
   render: function () {
     var showHomeLink = true;
+    var {ChattyStore, UserStore} = this.state;
     return (<div>
       <StatusBar 
         showHomeLink={showHomeLink}
-        username={this.state.UserStore.username} 
-        connected={this.state.ChattyStore.connected}
-        showNewThreadBox={this.state.ChattyStore.showNewThreadBox}
-        unseenReplies={this.state.UserStore.unseenReplies}
+        username={UserStore.username} 
+        connected={ChattyStore.connected}
+        showNewThreadBox={ChattyStore.showNewThreadBox}
+        unseenReplies={UserStore.unseenReplies}
+        fullRefresh={ChattyActions.fullRefresh}
+        reorderThreads={ChattyActions.reorderThreads}
+        showNewThread={ChattyActions.showNewThread}
+        cancelNewThread={ChattyActions.cancelNewThread}
+        
         />
-      <StatusMenu isMenuOpened={this.state.UserStore.isMenuOpened}
-        lastEventId={this.state.ChattyStore.eventId} 
-        username={this.state.UserStore.username} 
-        showLogin={this.state.UserStore.showLogin}
-        loginMessage={this.state.UserStore.loginMessage}
+      <StatusMenu isMenuOpened={UserStore.isMenuOpened}
+        lastEventId={ChattyStore.eventId} 
+        username={UserStore.username} 
+        showLogin={UserStore.showLogin}
+        loginMessage={UserStore.loginMessage}
         showSearch={this.state.showSearch}
-        pms={this.state.UserStore.pms}
-        unreadPMs={this.state.UserStore.unreadPMs}
-        unseenReplies={this.state.UserStore.unseenReplies}
-        totalPMs={this.state.UserStore.totalPMs}
+        pms={UserStore.pms}
+        unreadPMs={UserStore.unreadPMs}
+        unseenReplies={UserStore.unseenReplies}
+        totalPMs={UserStore.totalPMs}
+        reorderThreads={ChattyActions.reorderThreads}
+        showThreads={ChattyActions.showThreads}
         />
-      <RouteHandler UserStore={this.state.UserStore} 
-        ChattyStore={this.state.ChattyStore} />
+        {React.cloneElement(this.props.children, {UserStore,ChattyStore,chattyActions: ChattyActions,userActions: UserActions})}
     </div>);
   }
 });
 
 var routes = (
   <Route handler={App}>
-    <DefaultRoute name="ChattyHome" handler={ChattyHome} />
-    <Route name="SinglePost" handler={SinglePost}/>
+    <Route path="/" component={App} >
+      <IndexRoute component={ChattyHome} />
+      <Route path="SinglePost/:threadId" component={SinglePost}/>
+    </Route>
+    
   </Route>
 );
 
-Router.run(routes, function (Handler) {
-  React.render(<Handler/>, document.body);
-});
+ReactDOM.render(<Router>{routes}</Router>, document.getElementById('app'));
