@@ -1,9 +1,7 @@
-var React = require("react/addons");
+var React = require("react");
 var Router = require('react-router');
 var { Route, DefaultRoute, RouteHandler, Link } = Router;
-var ChattyActions = require("../store/chattyactions.js");
 var ParentComment = require("../view/posts/parentcomment.js");
-var _ = require("lodash");
 
 var styles = {
   commentBody : {
@@ -11,7 +9,6 @@ var styles = {
     paddingLeft: 20,
     paddingRight: 20,
     fontFamily: 'Helvetica,Arial,sans-serif',
-    fontSize: '0.8em',
   },
 };
 
@@ -27,26 +24,26 @@ var SinglePost =  React.createClass({
     };
   },
   componentWillReceiveProps: function(nextProps) {
-    var { router } = this.context;
-    var threadId = router.getCurrentQuery().threadId;
-    var postId = router.getCurrentQuery().postId;
+    var threadId = this.props.routeParams.threadId;
+    //var postId = router.getCurrentQuery().postId;
     if(this.state.threadId != threadId) {
       this.setState({
         threadId: Number(threadId),
-        postId: Number(postId),
+        postId: Number(0),
         focused: false,
       });
-      ChattyActions.loadPostIfNotFound(threadId);
+      this.props.loadPostIfNotFound(threadId);
     } else if(!this.state.focused) {
         var t = this.props.ChattyStore.findThreadByPostId(threadId);
         if (t) {
-          ChattyActions.selectComment(t.id,this.state.postId);
+          this.props.selectComment(t.id,this.state.postId);
           this.setState({focused: true});
         }
       }
   },
   render: function () {
     if(!this.state.threadId) return null;
+    
     var comment = this.props.ChattyStore.findThreadByPostId(this.state.threadId);
     if(!comment) { 
       return null;
